code/webapp/requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.13
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
Django==2.1
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.7
importlib-metadata==1.5.0
jmespath==0.9.4
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
numpy==1.18.1
opencv-python==4.2.0.32
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.0.0
protobuf==3.11.3
pycparser==2.19
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.11.2
zipp==2.2.0
